API enhancements - get_reference_target for elements, get ..._value for character_data

+++ autosar-data/src/lib.rs
@@ -160,7 +160,7 @@ pub enum ElementContent {
 pub enum CharacterData {
     Enum(EnumItem),
     String(String),
-    UnsignedInteger(usize),
+    UnsignedInteger(u64),
     Double(f64),
 }
